requirements.txt
colorama==0.4.6
langchain
langchain_community
langchain_experimental
langchain_openai
pandas
pydantic
python-dotenv
chromadb
unstructured
unstructured[pdf]
unstructured[xlsx]
langchain_text_splitters
sentence-transformers

